+++ DotNetEcuador.API/Common/Constants.cs
