+++ DotNetEcuador.API/Controllers/AreaOfInterestController.cs
