+++ DotNetEcuador.API/Controllers/CommunityController.cs
