+++ DotNetEcuador.API/Infraestructure/Repositories/Repository.cs
