+++ DotNetEcuador.API/Infraestructure/Services/CommunityService.cs
