+++ DotNetEcuador.API/Infraestructure/Services/IAreaOfInterestService.cs
