+++ DotNetEcuador.API/Models/AreaOfInterest.cs
