+++ DotNetEcuador.API/Models/CommunityMember.cs
