+++ DotNetEcuador.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.11.35312.102
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetEcuador.API", "api\DotNetEcuador.API.csproj", "{8F6D3F3D-1CCF-4E18-BC9C-67DF10364FB0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DotNetEcuador.API", "DotNetEcuador.API\DotNetEcuador.API.csproj", "{8F6D3F3D-1CCF-4E18-BC9C-67DF10364FB0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DotNetEcuador.Tests", "DotNetEcuador.Tests\DotNetEcuador.Tests.csproj", "{2DFD79E8-5349-4801-A429-DEAB65F26A50}"
 EndProject
