--- api/.env.example
@@ -1,2 +1,0 @@
-﻿MONGO_CONNECTION_STRING=mongodb://localhost:27017/
-MONGO_DATABASE=MiBaseDeDatos
